+++ pages/page/[page].tsx
@@ -3,20 +3,9 @@ import Head from "next/head";
 import Layout, { siteTitle } from "../../components/layout";
 import { getPaginatedPostData, getAllPosts, Post, MAX_PAGE_ENTRY } from "../../lib/posts";
 import utilStyles from "../../styles/utils.module.css";
-import { ContentBody } from "../../components/ContentBody";
-import { Indicator } from "../../components/post/pagenation/indicator";
+import { PostList, PostListProps } from "../../components/post/PostList";
 
-type Props = {
-  posts: Post[];
-  totalPages: number;
-  currentPage: number;
-};
-
-export default function Home({ posts, totalPages, currentPage }: Props) {
-  // publishだがindexには動線を表示させない記事。直接記事ページには行ける。
-  const EXPECT_SLUGS = ["test"];
-  //console.log(posts);
-
+export default function Home({ posts, totalPages, currentPage }: PostListProps) {
   return (
     <Layout home>
       <Head>
@@ -26,27 +15,7 @@ export default function Home({ posts, totalPages, currentPage }: Props) {
         <p>ゆるくやっていきます</p>
       </section>
 
-      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
-        <ul className={utilStyles.list}>
-          {posts &&
-            posts.map((post) => {
-              if (EXPECT_SLUGS.includes(post.slug)) {
-                return <></>;
-              }
-              return (
-                <li className={utilStyles.listItem} key={post.slug}>
-                  <ContentBody post={post} />
-                </li>
-              );
-            })}
-        </ul>
-      </section>
-      <Indicator
-        totalPages={totalPages}
-        currentPage={currentPage}
-        prevDisabled={currentPage === 1}
-        nextDisabled={currentPage === totalPages}
-      />
+      <PostList posts={posts} totalPages={totalPages} currentPage={currentPage} />
     </Layout>
   );
 }
